WIP updating payment API
