+++ api/paymentRequest.js
@@ -1,35 +1,7 @@
 /* eslint-env node */
-const headers = { 'Content-Type': 'application/json' };
 const methods = ['GET'];
-const stripeRate = 0.029;
-const stripeFlatCharge = 0.3;
-const taxRate = 0.0725;
-const currency = 'USD';
-
-function toCurrency(num) {
-	return parseFloat(num.toFixed(2));
-}
-
-function getTotal({ displayItems = [], modifiers: { additionalDisplayItems = [] } = {}}) {
-	const total = [...displayItems, ...additionalDisplayItems]
-		.reduce((total, { amount: { value = 0 } = {}}) => total + value, 0);
-
-	return toCurrency(total);
-}
-
-async function calculateCardFee(req) {
-	const total = getTotal(req.details);
-	const fee = toCurrency((total * stripeRate) + stripeFlatCharge);
-	return fee;
-}
-
-async function calculateTaxes(req) {
-	return toCurrency(getTotal(req) * taxRate);
-}
-
-async function calculateShipping() {
-	return 0;
-}
+const { calculateShipping, calculateCardFee, calculateTaxes, getTotal } = require('./stripe-utils.js');
+const { headers, currency } = require('./stripe-consts.js');
 
 exports.handler = async function(event) {
 	switch(event.httpMethod) {
