+++ api/stripe-consts.js
@@ -1,0 +1,6 @@
+/* eslint-env node */
+export const headers = { 'Content-Type': 'application/json' };
+export const stripeRate = 0.029;
+export const stripeFlatCharge = 0.3;
+export const taxRate = 0.0725;
+export const currency = 'USD';
